fix: more difficulty with eslint

+++ packages/base/scripts/bootstrap/bootstrapConfigFiles.cjs
@@ -4,9 +4,11 @@ const getPluginsObject = require('../../util/getPluginsObject.cjs')
 async function bootstrapConfigFiles() {
   const files = await getPluginsObject('getConfigFiles')
   for (const file in files) {
-    await fs.writeFile(file, files[file], 'utf8')
-    // eslint-disable-next-line no-console
-    console.error(`wrote ${file}`)
+    if (!(await fs.pathExists(file))) {
+      await fs.writeFile(file, files[file], 'utf8')
+      // eslint-disable-next-line no-console
+      console.error(`wrote ${file}`)
+    }
   }
 }
